src/backend/map_generator/requirements.txt
opensimplex
pillow
numpy
oxl-utils